+++ CHANGELOG.md
@@ -5,10 +5,14 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [Unreleased](https://github.com/alexdlaird/hookee/compare/0.0.6...HEAD)
 ### Added
-- Additional `ngrok` passthrough args to the CLI: `region`, `subdomain`, and `auth`.
-- `--last_request` as an arg to CLI, which can point to a Python script, where `last_request.run(request)` will be called after all plugins have processed a request to the default `/webhook`.").
-- `--last_response` as an arg to CLI, which can point to a Python script, where `last_response.run(request, response)` will be called after all plugins have generated the default `/webhook`'s response..
+- Additional `ngrok` passthrough args: `region`, `subdomain`, and `auth`.
+- `list-plugins` and `enabled-plugins` to the CLI.
+- `--request` to the CLI, a Python script whose `run(request)` method will be called by the default `/webhook` after all request plugins have run.
+- `--response` to the CLI, s Python script whose `run(request, response)` method will be called by the default `/webhook` after all response plugins have run.
 - Better console output.
+
+### Changed
+- Combined config-related CLI commands into single command, `set-config KEY VALUE`.
 
 ### Fixed
 - Stability improvements.